+++ src/main/java/roomescape/payment/service/TossPaymentRestClient.java
@@ -1,4 +1,4 @@
-package roomescape.payment.infrastructure;
+package roomescape.payment.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Value;
